refactor: 인기테마 조회 기능 유스케이스로 분리
