+++ src/main/java/roomescape/service/theme/ThemeService.java
@@ -1,7 +1,6 @@
 package roomescape.service.theme;
 
 import java.util.List;
-import roomescape.dto.theme.PopularThemeResponse;
 import roomescape.dto.theme.ThemeRequest;
 import roomescape.dto.theme.ThemeResponse;
 
@@ -11,6 +10,4 @@ public interface ThemeService {
     List<ThemeResponse> getAll();
 
     void deleteById(Long id);
-
-    List<PopularThemeResponse> getPopularThemes();
 }
