+++ src/main/java/roomescape/service/theme/popular/PopularThemeService.java
@@ -1,0 +1,37 @@
+package roomescape.service.theme.popular;
+
+import java.time.LocalDate;
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.dto.theme.PopularThemeResponse;
+import roomescape.entity.ThemeEntity;
+import roomescape.exception.theme.ThemeNotFoundException;
+import roomescape.repository.reservation.ReservationRepository;
+import roomescape.repository.theme.ThemeRepository;
+
+@Service
+public class PopularThemeService implements PopularThemeUseCase {
+    private final ThemeRepository themeRepository;
+    private final ReservationRepository reservationRepository;
+
+    public PopularThemeService(ThemeRepository themeRepository, ReservationRepository reservationRepository) {
+        this.themeRepository = themeRepository;
+        this.reservationRepository = reservationRepository;
+    }
+
+    @Override
+    public List<PopularThemeResponse> getPopularThemes() {
+        LocalDate endDate = LocalDate.now().minusDays(1);
+        LocalDate startDate = endDate.minusDays(7);
+        List<Long> themeIds = reservationRepository.findThemeIdsOrderByReservationCountBetween(startDate, endDate)
+                .stream()
+                .limit(10)
+                .toList();
+
+        return themeIds.stream().map(themeId -> {
+            ThemeEntity themeEntity = themeRepository.findById(themeId)
+                    .orElseThrow(ThemeNotFoundException::new);
+            return PopularThemeResponse.of(themeEntity);
+        }).toList();
+    }
+}
