+++ src/main/java/roomescape/service/theme/popular/PopularThemeUseCase.java
@@ -1,0 +1,8 @@
+package roomescape.service.theme.popular;
+
+import java.util.List;
+import roomescape.dto.theme.PopularThemeResponse;
+
+public interface PopularThemeUseCase {
+    List<PopularThemeResponse> getPopularThemes();
+}
